cleanup

+++ source/cServerData.cpp
@@ -713,7 +713,7 @@ auto CServerData::ResetDefaults() -> void
 	CombatBloodEffectChance( 75 );
 	GlobalAttackSpeed( 1.0 );
 	NPCSpellCastSpeed( 1.0 );
-	FishingStaminaLoss( 2.0 );
+	FishingStaminaLoss( static_cast<std::uint16_t>(2.0) );
 	CombatArmorClassDamageBonus( false );
 	AlchemyDamageBonusEnabled( false );
 	AlchemyDamageBonusModifier( 5 );
